kinethos_cdk/services/telegram_bot/requirements.txt
python-telegram-bot==21.*
python-dotenv>=1.0.1